move majors to config
