+++ client/src/config/majors.js
